+++ app/(tabs)/favorite.tsx
@@ -1,0 +1,17 @@
+import { Text, View, StyleSheet } from "react-native";
+
+export default function FavoriteScreen() {
+  return (
+    <View style={styles.container}>
+      <Text>즐겨찾기</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+  },
+});
