--- app/favorite.tsx
@@ -1,15 +1,0 @@
-import { View, Text } from "react-native";
-
-export default function FavoriteScreen() {
-  return (
-    <View
-      style={{
-        flex: 1,
-        justifyContent: "center",
-        alignItems: "center",
-      }}
-    >
-      <Text>즐겨찾기</Text>
-    </View>
-  );
-}
