--- app/index.tsx
@@ -1,15 +1,0 @@
-import { Text, View } from "react-native";
-
-export default function Index() {
-  return (
-    <View
-      style={{
-        flex: 1,
-        justifyContent: "center",
-        alignItems: "center",
-      }}
-    >
-      <Text>지도</Text>
-    </View>
-  );
-}
