--- app/my.tsx
@@ -1,15 +1,0 @@
-import { Text, View } from "react-native";
-
-export default function MyScreen() {
-  return (
-    <View
-      style={{
-        flex: 1,
-        justifyContent: "center",
-        alignItems: "center",
-      }}
-    >
-      <Text>내 정보</Text>
-    </View>
-  );
-}
