--- app/share.tsx
@@ -1,15 +1,0 @@
-import { Text, View } from "react-native";
-
-export default function SharedScreen() {
-  return (
-    <View
-      style={{
-        flex: 1,
-        justifyContent: "center",
-        alignItems: "center",
-      }}
-    >
-      <Text>공유</Text>
-    </View>
-  );
-}
